support async inserts
performance seems to be much faster if you have a lot of inserts

based on code i wrote for influxdb/api/graphite/api.go (commit
856a94cc78eff2b633fbee388339231354daaf89)
i made the insertHandler aware of the async setting, instead of handling
it in the committer,
so that we can easily execute inserts one way or another.
i.e. we can implement having async option globally enabled, but still
prefix queries with 'sync' to execute those synchronized
this makes the loop also simple, it doesn't need to be aware of two
kinds of inserts.